# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Supply Chain OSINT AI API
  version: 1.0.0-draft
  description: |
    Contract between the frontend demo site and the backend. See API_SPEC.md for
    semantics, SSE event schemas, and the agent protocol. Field names and types
    here are authoritative.
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /uploads:
    post:
      tags: [uploads]
      operationId: createUpload
      summary: Upload a CSV bill of materials
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Upload' } } } }
        '400': { $ref: '#/components/responses/Error' }
  /uploads/{upload_id}:
    get:
      tags: [uploads]
      operationId: getUpload
      parameters: [{ $ref: '#/components/parameters/UploadId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Upload' } } } }
        '404': { $ref: '#/components/responses/Error' }

  # ------------------------------------------------------------ runs
  /runs:
    post:
      tags: [runs]
      operationId: createRun
      summary: Start a research run (or replay a cached one)
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunCreate' }
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
    get:
      tags: [runs]
      operationId: listRuns
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/RunStatus' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Run' } }
                  next_cursor: { type: [string, 'null'] }
  /runs/{run_id}:
    get:
      tags: [runs]
      operationId: getRun
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /runs/{run_id}/answers:
    post:
      tags: [runs]
      operationId: answerRunQuestion
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question_id, choice]
              properties:
                question_id: { type: string }
                choice: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /runs/{run_id}/cancel:
    post:
      tags: [runs]
      operationId: cancelRun
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
  /runs/{run_id}/events:
    get:
      tags: [runs]
      operationId: streamRunEvents
      summary: Server-sent events for a run (see API_SPEC.md §3.1)
      parameters:
        - { $ref: '#/components/parameters/RunId' }
        - name: Last-Event-ID
          in: header
          schema: { type: string }
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema: { $ref: '#/components/schemas/RunEvent' }

  # ------------------------------------------------------------ graphs
  /graphs:
    get:
      tags: [graphs]
      operationId: listGraphs
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/GraphMeta' } }
                  next_cursor: { type: [string, 'null'] }
  /graphs/{graph_id}:
    get:
      tags: [graphs]
      operationId: getGraph
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { $ref: '#/components/parameters/Revision' }
        - name: include
          in: query
          description: 'Comma-separated. Supported: claims'
          schema: { type: string }
        - name: tier_max
          in: query
          schema: { type: integer, minimum: 0 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Graph' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /graphs/{graph_id}/nodes/{node_id}:
    get:
      tags: [graphs]
      operationId: getNode
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { $ref: '#/components/parameters/NodeId' }
        - { $ref: '#/components/parameters/Revision' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/NodeDetail' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /graphs/{graph_id}/edges/{edge_id}:
    get:
      tags: [graphs]
      operationId: getEdge
      summary: Full edge with claims and evidence (edge inspector)
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { $ref: '#/components/parameters/EdgeId' }
        - { $ref: '#/components/parameters/Revision' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/EdgeDetail' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /graphs/{graph_id}/fork:
    post:
      tags: [graphs]
      operationId: forkGraph
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                revision: { type: integer }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/GraphMeta' } } } }
  /graphs/{graph_id}/mutations:
    post:
      tags: [graphs]
      operationId: applyMutations
      summary: Apply an atomic mutation batch
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - name: If-Match
          in: header
          required: true
          description: 'Current graph revision'
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MutationBatch' }
      responses:
        '200': { description: Applied, content: { application/json: { schema: { $ref: '#/components/schemas/MutationResult' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
    get:
      tags: [graphs]
      operationId: listMutations
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/MutationRecord' } }
                  next_cursor: { type: [string, 'null'] }
  /graphs/{graph_id}/export:
    get:
      tags: [graphs]
      operationId: exportGraph
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { $ref: '#/components/parameters/Revision' }
      responses:
        '200': { description: 'Full JSON export', content: { application/json: { schema: { $ref: '#/components/schemas/GraphExport' } } } }
  /graphs/{graph_id}/diff:
    get:
      tags: [graphs]
      operationId: diffGraph
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { name: from, in: query, required: true, schema: { type: integer } }
        - { name: to, in: query, required: true, schema: { type: integer } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GraphDiff' } } } }

  # ------------------------------------------------------------ claims / sources
  /claims/{claim_id}:
    get:
      tags: [claims]
      operationId: getClaim
      parameters: [{ $ref: '#/components/parameters/ClaimId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Claim' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /claims/{claim_id}/review:
    post:
      tags: [claims]
      operationId: reviewClaim
      parameters: [{ $ref: '#/components/parameters/ClaimId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verdict]
              properties:
                verdict: { type: string, enum: [accept, reject, dispute] }
                note: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Claim' } } } }
  /sources/{source_id}:
    get:
      tags: [claims]
      operationId: getSource
      parameters: [{ name: source_id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Source' } } } }

  # ------------------------------------------------------------ enrichments
  /graphs/{graph_id}/enrichments:
    post:
      tags: [enrichments]
      operationId: createEnrichment
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kinds]
              properties:
                kinds:
                  type: array
                  minItems: 1
                  items: { type: string, enum: [geography, concentration, market_exposure] }
                node_ids: { type: array, items: { type: string } }
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Enrichment' } } } }
  /graphs/{graph_id}/enrichments/{enrichment_id}:
    get:
      tags: [enrichments]
      operationId: getEnrichment
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { name: enrichment_id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Enrichment' } } } }

  # ------------------------------------------------------------ geography
  /graphs/{graph_id}/geography:
    get:
      tags: [geography]
      operationId: getGraphGeography
      summary: GeoJSON FeatureCollection of facilities and material-producing countries
      parameters:
        - { $ref: '#/components/parameters/GraphId' }
        - { $ref: '#/components/parameters/Revision' }
      responses:
        '200': { description: OK, content: { application/geo+json: { schema: { $ref: '#/components/schemas/GeoFeatureCollection' } } } }
  /materials/{material_node_id}/production:
    get:
      tags: [geography]
      operationId: getMaterialProduction
      parameters:
        - { name: material_node_id, in: path, required: true, schema: { type: string } }
        - { name: year, in: query, schema: { type: integer } }
        - { name: stage, in: query, schema: { type: string, enum: [mine, refinery, export] } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ProductionShares' } } } }
        '404': { $ref: '#/components/responses/Error' }
...
  /bom/{bom_id}:
    get:
      tags: [bom]
      operationId: getBillOfMaterialsEstimate
      summary: A persisted BOM estimate
      parameters:
        - { name: bom_id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: './bom-estimate.schema.json' } } } }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    IdempotencyKey: { name: Idempotency-Key, in: header, required: false, schema: { type: string, maxLength: 128 } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
    Revision: { name: revision, in: query, description: 'Graph revision; omit for latest', schema: { type: integer, minimum: 0 } }
    UploadId: { name: upload_id, in: path, required: true, schema: { type: string } }
    RunId: { name: run_id, in: path, required: true, schema: { type: string } }
    GraphId: { name: graph_id, in: path, required: true, schema: { type: string } }
    NodeId: { name: node_id, in: path, required: true, schema: { type: string } }
    EdgeId: { name: edge_id, in: path, required: true, schema: { type: string } }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    Upload:
      type: object
      required: [id, filename, row_count, columns, created_at]
      properties:
        id: { type: string }
        filename: { type: string }
        row_count: { type: integer }
        columns: { type: array, items: { type: string } }
        preview: { type: array, items: { type: object, additionalProperties: true }, maxItems: 10 }
        warnings: { type: array, items: { type: string } }
        content_hash: { type: string }
        created_at: { type: string, format: date-time }

    # ---------------- runs
    RunStatus:
      type: string
      enum: [queued, awaiting_input, running, completed, partial, failed, cancelled]
...